backend-django/requirements.txt
asgiref==3.5.2
backcall==0.2.0
colorama==0.4.5
decorator==5.1.1
Django==3.2.12
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ipython==7.34.0
jedi==0.18.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.29
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.30
protobuf==4.21.4
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
traitlets==5.3.0
typing-extensions==4.3.0
wcwidth==0.2.5
